+++ packages/react-static/src/utils/binHelper.js
@@ -11,15 +11,42 @@ let ignorePath
 
 const originalRequire = Module.prototype.require
 
+// Check and see if we are running react-static from the repo
+const needsWorkspaceCheck = __dirname.includes(
+  '/react-static/packages/react-static/'
+)
+
+// Recursively checks a module to see if it originated from a
+// react-static package in the repo
+const inRepo = mod => {
+  if (
+    !mod.filename.includes('react-static/packages/react-static/') &&
+    mod.filename.includes('react-static/packages/')
+  ) {
+    return true
+  }
+  if (mod.parent) {
+    return inRepo(mod.parent)
+  }
+  return false
+}
+
 // The following ensures that there is always only a single (and same)
 // copy of React in an app at any given moment.
 // eslint-disable-next-line
 Module.prototype.require = function(modulePath) {
+  // If we are running in the repo, we need to make sure
+  // module resolutions coming from other react-static packages
+  // are first attempted from the
+  const isInWorkspace = needsWorkspaceCheck && inRepo(this)
+
+  // Only redirect resolutions to non-relative and non-absolute modules
   if (!modulePath.startsWith('.') && !modulePath.startsWith('/')) {
     if (
-      ['react', 'react-dom', 'styled-components'].some(d =>
-        modulePath.includes(d)
-      )
+      // If module is in the repo try and redirect
+      isInWorkspace ||
+      // Always try and redirect react and react-dom resolutions
+      ['react', 'react-dom'].some(d => modulePath.includes(d))
     ) {
       try {
         modulePath = resolveFrom(