+++ packages/react-static/templates/basic/src/App.js
@@ -1,11 +1,13 @@
 import React from 'react'
 import { Root, Routes, addPrefetchExcludes } from 'react-static'
 //
-import { Link } from 'components/Router'
+import { Link, Router } from 'components/Router'
+import Dynamic from 'containers/Dynamic'
 
 import './app.css'
 
-addPrefetchExcludes(['test'])
+// Any routes that start with 'dynamic' will be treated as non-static routes
+addPrefetchExcludes(['dynamic'])
 
 function App() {
   return (
@@ -14,11 +16,14 @@ function App() {
         <Link to="/">Home</Link>
         <Link to="/about">About</Link>
         <Link to="/blog">Blog</Link>
-        <Link to="/test">Test</Link>
+        <Link to="/dynamic">Dynamic</Link>
       </nav>
       <div className="content">
         <React.Suspense fallback={<em>Loading...</em>}>
-          <Routes />
+          <Router>
+            <Dynamic path="dynamic" />
+            <Routes path="*" />
+          </Router>
         </React.Suspense>
       </div>
     </Root>