+++ packages/react-static/templates/basic/src/components/Router.js
@@ -1,8 +1,1 @@
-// react-location
-// export { Link } from 'react-location'
-
-// @reach/router
-// export { Link } from '@reach/router'
-
-// react-router
-export { Link } from 'react-router-dom'
+export { Link, Router } from '@reach/router'