+++ packages/react-static/templates/basic/src/pages/index.js
@@ -1,12 +1,7 @@
 import React from 'react'
-import styled from '@emotion/styled'
-
-const Div = styled.div`
-  color: blue;
-`
 
 export default () => (
-  <Div style={{ textAlign: 'center' }}>
+  <div style={{ textAlign: 'center' }}>
     <h1>Welcome to React-Static</h1>
-  </Div>
+  </div>
 )